+++ earcrate/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.7.1"
+__version__ = "0.7.7"
 from earcrate.cli import main